+++ javascript/src/ParameterType.ts
@@ -49,6 +49,7 @@ export default class ParameterType<T> {
    * @param transform {Function} function transforming string to another type. May be null.
    * @param useForSnippets {boolean} true if this should be used for snippets. Defaults to true.
    * @param preferForRegexpMatch {boolean} true if this is a preferential type. Defaults to false.
+   * @param builtin whether or not this is a built-in type
    */
   constructor(
     public readonly name: string | undefined,
@@ -56,7 +57,8 @@ export default class ParameterType<T> {
     public readonly type: Constructor<T> | Factory<T> | null,
     transform?: (...match: string[]) => T | PromiseLike<T>,
     public readonly useForSnippets?: boolean,
-    public readonly preferForRegexpMatch?: boolean
+    public readonly preferForRegexpMatch?: boolean,
+    public readonly builtin?: boolean
   ) {
     if (transform === undefined) {
       transform = (s) => s as unknown as T